Refined "How to Use the Kit" and "Accessible"
I removed the disambiguation between "Bocoupers" and "everyone else" - seeing that this kit is open, we should really try to write it in a way that speaks to any potential user. In this sense, everyone is a Bocouper. 

I added the word "abilities" to our statement on accessibility because the word accessible connotes that it is [a11y compliant](https://wiki.openoffice.org/wiki/Accessibility_(A11y)_Quick_Test_Check_List). We might even want to say that outright.